feat(modal): add resetRoot and withResetRoot to a new hook file, apply only once in modal code
