+++ packages/storybook/stories/hook.ts
@@ -1,0 +1,15 @@
+const resetRoot = () => {
+  const storybookRoot = document.querySelector('#storybook-root');
+  if (storybookRoot) {
+    storybookRoot.innerHTML = '';
+  }
+};
+
+const withResetRoot = (storyFn: any) => {
+  resetRoot();
+  return storyFn();
+};
+export {
+  resetRoot,
+  withResetRoot,
+};
